requirements.txt
crewai
crewai_tools
python-binance
pandas
ta
openai
langchain
pydantic
PyYAML
python-dotenv